Several improvements and fixes: - Output 1 message when using /delete /pokemon and /list instead of 1 for each gym. - Change abort key to delete gym when submitted via location and not saved yet. - Removed /new as we don't need it anymore - Minor improvements and fixes

+++ mods/exit.php
@@ -18,8 +18,14 @@ answerCallbackQuery($update['callback_query']['id'], $msg);
 // Edit the message.
 edit_message($update, $msg, $keys);
 
+// Set gym_user_id tag.
+$gym_user_id = '#' . $update['callback_query']['from']['id'];
+
+// Get gym.
+$gym = get_gym($data['id']);
+
 // Delete gym from database.
-if($data['arg'] == 2) {
+if($gym['gym_name'] == $gym_user_id && $gym['show_gym'] == 0 && $data['arg'] == 2) {
     delete_gym($data['id']);
 }
